$ Refactor 2 prompts to use shared component code
