--- src/renderer/components/ft-playlist-select-quick-bookmark-target-prompt/ft-playlist-select-quick-bookmark-target-prompt.css
@@ -1,53 +1,0 @@
-.heading {
-  text-align: center;
-}
-
-.searchInputsRow {
-  display: grid;
-
-  /* 2 columns */
-  grid-template-columns: 1fr auto;
-  column-gap: 16px;
-}
-
-@media only screen and (width <= 800px) {
-  .searchInputsRow {
-    /* Switch to 2 rows from 2 columns */
-    grid-template-columns: auto;
-    grid-template-rows: auto auto;
-  }
-}
-
-.optionsRow {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-template-rows: 1fr;
-  align-items: center;
-}
-
-@media only screen and (width <= 800px) {
-  .optionsRow {
-    /* Switch to 2 rows from 2 columns */
-    grid-template-columns: auto;
-    grid-template-rows: auto auto;
-    align-items: stretch;
-  }
-}
-
-.sortSelect {
-  /* Put it on the right */
-  margin-inline-start: auto;
-}
-
-.playlists-container {
-  box-shadow: inset 0 0 5px rgb(0 0 0 / 50%);
-
-  /* Use remaining height */
-  flex-grow: 1;
-  overflow-y: scroll;
-}
-
-.playlist-selector-container {
-  /* Make them look selectable */
-  cursor: pointer;
-}
