+++ src/renderer/components/ft-playlist-selection-prompt/ft-playlist-selection-prompt.css
